Added the observation from Rúbens.

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: asepulve <asepulve@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 14:20:47 by asepulve          #+#    #+#             */
-/*   Updated: 2023/07/19 23:03:49 by asepulve         ###   ########.fr       */
+/*   Updated: 2023/07/20 00:06:20 by asepulve         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@
 
 typedef struct s_philo
 {
-	int				n_philos;
-	int				time_to_die;
-	int				time_to_eat;
-	int				time_to_sleep;
-	int				n_times_must_eat;
+	long			n_philos;
+	long			time_to_die;
+	long			time_to_eat;
+	long			time_to_sleep;
+	long			n_times_must_eat;
 	pthread_t		*philos;
 	pthread_mutex_t	*forks;
 }	t_philo;
